+++ foo_spider_monkey_panel/utils/winapi_error_helpers.cpp
@@ -17,12 +17,23 @@ std::u8string MessageFromErrorCode( DWORD errorCode )
 
 void ThrowParsedWinapiError( DWORD errorCode, std::string_view functionName )
 {
+    const auto errorMessage = [errorCode]() -> std::u8string {
+        if (errorCode == ERROR_SUCCESS)
+        {// some functions are bugged, e.g. CreateFont (<https://github.com/TheQwertiest/foo_spider_monkey_panel/issues/92>)
+            return "Function failed, but returned a `SUCCESS` error code, which is usually caused by a bugged WinAPI. "
+                   "One such case is when process runs out of GDI handles and can't create a new GDI object.";
+        }
+        else
+        {
+            return MessageFromErrorCode( errorCode );
+        }
+    }();
     throw SmpException( fmt::format( "WinAPI error:\n"
                                      "  {} failed with error ({:#x}):\n"
                                      "    {}",
                                      functionName,
                                      errorCode,
-                                     MessageFromErrorCode( errorCode ) ) );
+                                     errorMessage ) );
 }
 
 } // namespace